Fix bug in self messages
Detected in errbit.
