+++ test/integration/desktop_messages_test.rb
@@ -1,8 +1,15 @@
 # frozen_string_literal: true
 require 'test_helper'
 require 'support/desktop_integration'
-require 'integration/concerns/messages'
+require 'integration/concerns/standard_messages'
+require 'integration/concerns/self_messages'
 
-class DesktopMessagesTest < DesktopIntegrationTest
-  include Messages
+module DesktopMessagesTest
+  class StandardMessagesTest < DesktopIntegrationTest
+    include StandardMessages
+  end
+
+  class SelfDesktopMessagesTest < DesktopIntegrationTest
+    include SelfMessages
+  end
 end
